pyproject.toml
[tool.poetry]
name = "web-hw-11"
version = "0.1.0"
description = ""
authors = ["Volodymyr Pryhlo <vova.pruglo22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
python-multipart = "^0.0.9"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
psycopg2 = "^2.9.9"
python-jose = "^3.3.0"
passlib = "^1.7.4"
alembic = "^1.13.1"
bcrypt = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
